+++ modules/credit_datasets.py
@@ -87,7 +87,7 @@ class TabularDataset(Dataset):
             df[self.continuous] /= self.std
             
             self.train = df
-            self.x_data = df.to_numpy()
+            self.x_data = df.to_numpy().astype(np.float32)
         else:
             self.train_raw = base.iloc[:split_num]
             self.test_raw = base.iloc[split_num:]
@@ -102,7 +102,7 @@ class TabularDataset(Dataset):
             df[self.continuous] /= self.std
             
             self.test = df
-            self.x_data = df.to_numpy()
+            self.x_data = df.to_numpy().astype(np.float32)
         
         # Output Information
         self.OutputInfo_list = []
